refactor(pipeline): move run logic to pipeline module
- Moved async run logic from main.rs to pipeline.rs
- Added pipeline::run for CSV fetch, parse, and output
- Updated cli.rs and main.rs to use pipeline::run
- Split record processing into helper functions
- Improved code organization and modularity

+++ src/media.rs
@@ -41,7 +41,7 @@ pub fn make_media_path(
     sheet: &str,
     id: &str,
     media_type: &crate::config::MediaType,
-    index: usize,
+    index: i32,
     url: &str,
 ) -> std::path::PathBuf {
     let ext = url
